Created explicit NRCreator and NRContributor data types (to be used from compchem)
