+++ packages/nr-metadata/nr_metadata/common/services/records/schema_common.py
@@ -6,7 +6,6 @@ from marshmallow.validate import OneOf
 from marshmallow_utils.fields import TrimmedString
 from oarepo_runtime.services.schema.i18n import I18nStrField, MultilingualField
 from oarepo_runtime.services.schema.marshmallow import BaseRecordSchema, DictOnlySchema
-from oarepo_runtime.services.schema.polymorphic import PolymorphicSchema
 from oarepo_runtime.services.schema.validation import (
     CachedMultilayerEDTFValidator,
     validate_date,
@@ -14,14 +13,12 @@ from oarepo_runtime.services.schema.validation import (
 
 from nr_metadata.common.services.records.schema_datatypes import (
     NRAccessRightsVocabularySchema,
-    NRAffiliationVocabularySchema,
-    NRContributorTypeVocabularySchema,
+    NRContributorSchema,
+    NRCreatorSchema,
     NREventSchema,
     NRFundingReferenceSchema,
     NRGeoLocationSchema,
     NRLanguageVocabularySchema,
-    NROrganizationSchema,
-    NRPersonSchema,
     NRRelatedItemSchema,
     NRResourceTypeVocabularySchema,
     NRRightsVocabularySchema,
@@ -31,8 +28,6 @@ from nr_metadata.common.services.records.schema_datatypes import (
 )
 from nr_metadata.schema.identifiers import (
     NRObjectIdentifierSchema,
-    NROrganizationIdentifierSchema,
-    NRPersonIdentifierSchema,
     NRSystemIdentifierSchema,
 )
 
@@ -119,17 +114,6 @@ class NRCommonMetadataSchema(Schema):
     version = ma_fields.String()
 
 
-class NRContributorSchema(PolymorphicSchema):
-    class Meta:
-        unknown = ma.RAISE
-
-    Organizational = ma_fields.Nested(lambda: NRContributorOrganizationSchema())
-
-    Personal = ma_fields.Nested(lambda: NRContributorPersonSchema())
-
-    type_field = "nameType"
-
-
 class AdditionalTitlesSchema(DictOnlySchema):
     class Meta:
         unknown = ma.RAISE
@@ -140,52 +124,3 @@ class AdditionalTitlesSchema(DictOnlySchema):
         required=True,
         validate=[OneOf(["translatedTitle", "alternativeTitle", "subtitle", "other"])],
     )
-
-
-class NRContributorOrganizationSchema(DictOnlySchema):
-    class Meta:
-        unknown = ma.RAISE
-
-    authorityIdentifiers = ma_fields.List(
-        ma_fields.Nested(lambda: NROrganizationIdentifierSchema())
-    )
-
-    contributorType = ma_fields.Nested(lambda: NRContributorTypeVocabularySchema())
-
-    fullName = ma_fields.String(required=True)
-
-    nameType = ma_fields.String(validate=[OneOf(["Organizational"])])
-
-
-class NRContributorPersonSchema(DictOnlySchema):
-    class Meta:
-        unknown = ma.RAISE
-
-    affiliations = ma_fields.List(
-        ma_fields.Nested(lambda: NRAffiliationVocabularySchema())
-    )
-
-    authorityIdentifiers = ma_fields.List(
-        ma_fields.Nested(lambda: NRPersonIdentifierSchema())
-    )
-
-    contributorType = ma_fields.Nested(lambda: NRContributorTypeVocabularySchema())
-
-    familyName = ma_fields.String()
-
-    fullName = ma_fields.String(required=True)
-
-    givenName = ma_fields.String()
-
-    nameType = ma_fields.String(validate=[OneOf(["Personal"])])
-
-
-class NRCreatorSchema(PolymorphicSchema):
-    class Meta:
-        unknown = ma.RAISE
-
-    Organizational = ma_fields.Nested(lambda: NROrganizationSchema())
-
-    Personal = ma_fields.Nested(lambda: NRPersonSchema())
-
-    type_field = "nameType"
