+++ packages/nr-metadata/nr_metadata/common/services/records/ui_schema_common.py
@@ -16,8 +16,8 @@ from oarepo_runtime.services.schema.ui import (
 
 from nr_metadata.common.services.records.ui_schema_datatypes import (
     NRAccessRightsVocabularyUISchema,
-    NRAffiliationVocabularyUISchema,
-    NRContributorTypeVocabularyUISchema,
+    NRContributorUISchema,
+    NRCreatorUISchema,
     NREventUISchema,
     NRFundingReferenceUISchema,
     NRGeoLocationUISchema,
@@ -31,8 +31,6 @@ from nr_metadata.common.services.records.ui_schema_datatypes import (
 )
 from nr_metadata.ui_schema.identifiers import (
     NRObjectIdentifierUISchema,
-    NROrganizationIdentifierUISchema,
-    NRPersonIdentifierUISchema,
     NRSystemIdentifierUISchema,
 )
 from nr_metadata.ui_schema.subjects import NRSubjectListField
@@ -126,85 +124,3 @@ class AdditionalTitlesUISchema(DictOnlySchema):
         required=True,
         validate=[OneOf(["translatedTitle", "alternativeTitle", "subtitle", "other"])],
     )
-
-
-class NRContributorOrganizationUISchema(DictOnlySchema):
-    class Meta:
-        unknown = ma.RAISE
-
-    authorityIdentifiers = ma_fields.List(
-        ma_fields.Nested(lambda: NROrganizationIdentifierUISchema())
-    )
-
-    contributorType = ma_fields.Nested(lambda: NRContributorTypeVocabularyUISchema())
-
-    fullName = ma_fields.String(required=True)
-
-    nameType = ma_fields.String(validate=[OneOf(["Organizational"])])
-
-
-class NRContributorPersonUISchema(DictOnlySchema):
-    class Meta:
-        unknown = ma.RAISE
-
-    affiliations = ma_fields.List(
-        ma_fields.Nested(lambda: NRAffiliationVocabularyUISchema())
-    )
-
-    authorityIdentifiers = ma_fields.List(
-        ma_fields.Nested(lambda: NRPersonIdentifierUISchema())
-    )
-
-    contributorType = ma_fields.Nested(lambda: NRContributorTypeVocabularyUISchema())
-
-    familyName = ma_fields.String()
-
-    fullName = ma_fields.String(required=True)
-
-    givenName = ma_fields.String()
-
-    nameType = ma_fields.String(validate=[OneOf(["Personal"])])
-
-
-class NRContributorUISchema(DictOnlySchema):
-    class Meta:
-        unknown = ma.RAISE
-
-    affiliations = ma_fields.List(
-        ma_fields.Nested(lambda: NRAffiliationVocabularyUISchema())
-    )
-
-    authorityIdentifiers = ma_fields.List(
-        ma_fields.Nested(lambda: NROrganizationIdentifierUISchema())
-    )
-
-    contributorType = ma_fields.Nested(lambda: NRContributorTypeVocabularyUISchema())
-
-    familyName = ma_fields.String()
-
-    fullName = ma_fields.String(required=True)
-
-    givenName = ma_fields.String()
-
-    nameType = ma_fields.String(validate=[OneOf(["Organizational"])])
-
-
-class NRCreatorUISchema(DictOnlySchema):
-    class Meta:
-        unknown = ma.RAISE
-
-    affiliations = ma_fields.List(
-        ma_fields.Nested(lambda: NRAffiliationVocabularyUISchema())
-    )
-
-    authorityIdentifiers = ma_fields.List(
-        ma_fields.Nested(lambda: NROrganizationIdentifierUISchema())
-    )
-
-    familyName = ma_fields.String()
-
-    fullName = ma_fields.String(required=True)
-
-    givenName = ma_fields.String()
-
-    nameType = ma_fields.String(validate=[OneOf(["Organizational"])])
