+++ packages/nr-metadata/nr_metadata/version.py
@@ -1,1 +1,1 @@
-__version__ = "2.0.19"
+__version__ = "2.0.20"
