+++ packages/oarepo-model-builder-nr/oarepo_model_builder_nr/version.py
@@ -1,1 +1,1 @@
-__version__ = "2.0.19"
+__version__ = "2.0.20"
